% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_data_ConsensusMouse}
\alias{load_data_ConsensusMouse}
\title{Load a consensus mouse component without conflating gene universes}
\usage{
load_data_ConsensusMouse(source = c("AIBS", "Macosko"), ...)
}
\arguments{
\item{source}{AIBS or Macosko.}

\item{...}{Further arguments to load_data.}
}
\value{
An abc_data list.
}
\description{
Load a consensus mouse component without conflating gene universes
}
