% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_data_HMBA_PatchSeq}
\alias{load_data_HMBA_PatchSeq}
\title{Load HMBA_PatchSeq metadata}
\usage{
load_data_HMBA_PatchSeq(download_base = "~/abc_atlas_cache", ...)
}
\arguments{
\item{download_base}{Cache directory outside the repository.}

\item{...}{Further arguments to load_data.}
}
\value{
An abc_data list.
}
\description{
Load HMBA_PatchSeq metadata
}
