% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_data_HMBA_spatial}
\alias{load_data_HMBA_spatial}
\title{Load HMBA MERSCOPE or Xenium spatial metadata}
\usage{
load_data_HMBA_spatial(species = c("Human", "Macaque", "Marmoset"), ...)
}
\arguments{
\item{species}{Human, Macaque or Marmoset.}

\item{...}{Further arguments to load_data.}
}
\value{
An abc_data list.
}
\description{
Load HMBA MERSCOPE or Xenium spatial metadata
}
