% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_data_SEAAD_Multiregion}
\alias{load_data_SEAAD_Multiregion}
\title{Load SEAAD_Multiregion metadata}
\usage{
load_data_SEAAD_Multiregion(download_base = "~/abc_atlas_cache", ...)
}
\arguments{
\item{download_base}{Cache directory outside the repository.}

\item{...}{Further arguments to load_data.}
}
\value{
An abc_data list.
}
\description{
Load SEAAD_Multiregion metadata
}
