% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_WHB.R, R/wrappers.R
\name{load_data_WHB}
\alias{load_data_WHB}
\title{Load ABC Atlas WHB data}
\usage{
load_data_WHB(download_base = "~/abc_atlas_cache", ...)

load_data_WHB(download_base = "~/abc_atlas_cache", ...)
}
\arguments{
\item{download_base}{Cache directory outside the repository.}

\item{...}{Further arguments to load_data.}
}
\value{
A list containing cell metadata and gene data.

An abc_data list.
}
\description{
This function loads the ABC Atlas WHB data from the specified download base directory.
}
